+++ pointapp/build.gradle
@@ -27,6 +27,17 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-security'
     implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
+
+    // Swagger
+    implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.0.2'
+
+    // modelMapper
+    implementation 'org.modelmapper:modelmapper:3.1.0'
+
+    // spring-boot-starter-validation
+    implementation 'org.springframework.boot:spring-boot-starter-validation:3.1.2'
+    
+
     runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5',
             'io.jsonwebtoken:jjwt-jackson:0.11.5'
 
@@ -36,11 +47,6 @@ dependencies {
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testImplementation 'org.springframework.security:spring-security-test'
 
-    // Swagger
-    implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.0.2'
-
-    // modelMapper
-    implementation 'org.modelmapper:modelmapper:3.1.0'
 
 }
