+++ pointapp/src/main/java/com/workjo/pointapp/auth/vo/LoginRequest.java
@@ -3,6 +3,7 @@ package com.workjo.pointapp.auth.vo;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import jakarta.validation.constraints.NotBlank;
 import lombok.*;
 
 
@@ -14,7 +15,9 @@ import lombok.*;
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class LoginRequest {
 
+	@NotBlank
 	private String loginId;
+	@NotBlank
 	private String password;
 
 }
